Made Integer class use BigInt for precision

+++ src/main/scala/algebra/structures/Integers.scala
@@ -2,7 +2,7 @@ package org.zetatypes.algebra.structures
 
 import org.zetatypes.algebra._
 
-case class Integer (x : Int) extends StandardLambdaRingElement[Integer] with PartialQAlgebraElement[Integer]{
+case class Integer (x : BigInt) extends StandardLambdaRingElement[Integer] with PartialQAlgebraElement[Integer]{
     override def +(that : Integer) = Integer(x + that.x) 
     override def *(that : Integer) = Integer(x * that.x)
     override def negation() = Integer(-x)
